+++ game/board.py
@@ -1,6 +1,7 @@
 from piezas.rook import Rook
 from piezas.king import King
 from piezas.horse import Horse
+from piezas.pawn import Pawn
 from game.exceptions import OutOfBoard
 
 
@@ -19,7 +20,7 @@ class Board:
         self.__positions__[0][7] = Rook("BLACK") # Black
         self.__positions__[7][7] = Rook("WHITE") # White
         self.__positions__[7][0] = Rook("WHITE") # White
-          #PISICIONES REY
+        #PISICIONES REY
         self.__positions__[0][4] = King("BLACK") 
         self.__positions__[7][4] = King("WHITE") 
 
@@ -27,6 +28,11 @@ class Board:
         self.__positions__[0][6] = Horse("BLACK") 
         self.__positions__[7][1] = Horse("WHITE") 
         self.__positions__[7][6] = Horse("WHITE") 
+
+        #POSICIONES PEONES
+        for col in range(8):
+            self.__positions__[6][col] = Pawn("White")
+            self.__positions__[1][col] = Pawn("Black")